partially mitigate setuptools-related problems
Since setuptools-49.2.0 the warning is issued if it finds distutils
already imported. This warning is set to being ignored, since it is
practically impossible to import setuptools all the time first (for
example, numpy imports distutils, which triggers the warning).

To make at least `import quantumsim` not lead to this warning, the
`__init__.py` for quantumsim is also fixed.